requirements.txt
# Core dependencies
transformers>=4.36.0  # For model handling
accelerate>=0.25.0    # For optimized inference
torch>=2.1.0          # PyTorch for ML operations
peft>=0.6.0           # Parameter-Efficient Fine-Tuning
bitsandbytes>=0.41.0  # Quantization support

# API Server
fastapi>=0.104.0      # API framework
uvicorn>=0.23.0       # ASGI server
pydantic>=2.4.0       # Data validation

# OpenRouter API
requests>=2.31.0      # HTTP client library
uuid>=1.30.0          # For session management

# Threat detection
beautifulsoup4>=4.12.0  # HTML parsing

# Configuration and environment
python-dotenv>=1.0.0    # Environment variable management

# Email/Reporting
aiosmtplib>=2.0.0      # Async SMTP client
jinja2>=3.1.0          # Template engine

# Data handling
pandas>=2.1.0          # Data analysis tools
